Add makefile. Divide and rename utils routines
Add makefile to facilitate execution and export of results.
