+++ utils/chronological.py
@@ -1,0 +1,7 @@
+from dateutil import parser
+
+
+def conform_date(date):
+   time = parser.parse(date)
+   formatted = time.strftime('%d/%m/%Y %H:%M:%S')
+   return formatted
